requirements.txt
asteval==1.0.1
contourpy==1.1.1
cycler==0.12.1
dill==0.3.8
fonttools==4.53.1
importlib_metadata==8.0.0
importlib_resources==6.4.0
json5==0.9.25
kiwisolver==1.4.5
lmfit==1.3.1
matplotlib==3.7.5
mpmath==1.3.0
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.1
pandas==2.0.3
pco==2.2.0
pillow==10.4.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
scipy==1.10.1
six==1.16.0
sympy==1.13.0
tzdata==2024.1
uncertainties==3.2.2
zipp==3.19.2
